# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "https://api.valerochkagym.tech",
      "description": "Production"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/admin/api/standard/{kind}/{id}": {
      "get": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "record_1",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogRecord"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "save",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StandardEdit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PushResult"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sync": {
      "get": {
        "tags": [
          "data-controller"
        ],
        "operationId": "snapshot",
        "parameters": [
          {
            "name": "X-Gym-Sync-Version",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Snapshot"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "data-controller"
        ],
        "operationId": "push",
        "parameters": [
          {
            "name": "X-Gym-Sync-Version",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PushResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/api/standard/{kind}/{id}/archive": {
      "post": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "archive",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StandardArchive"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PushResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/api/standard": {
      "get": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "list_1",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogPage"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "JsonNode": {
        "type": "object",
        "properties": {
          "pojo": {
            "type": "boolean"
          },
          "int": {
            "type": "boolean"
          },
          "long": {
            "type": "boolean"
          },
          "container": {
            "type": "boolean"
          },
          "nodeType": {
            "type": "string",
            "enum": [
              "ARRAY",
              "BINARY",
              "BOOLEAN",
              "MISSING",
              "NULL",
              "NUMBER",
              "OBJECT",
              "POJO",
              "STRING"
            ]
          },
          "string": {
            "type": "boolean"
          },
          "missingNode": {
            "type": "boolean"
          },
          "valueNode": {
            "type": "boolean"
          },
          "object": {
            "type": "boolean"
          },
          "number": {
            "type": "boolean"
          },
          "double": {
            "type": "boolean"
          },
          "bigDecimal": {
            "type": "boolean"
          },
          "bigInteger": {
            "type": "boolean"
          },
          "textual": {
            "type": "boolean",
            "deprecated": true
          },
          "boolean": {
            "type": "boolean"
          },
          "binary": {
            "type": "boolean"
          },
          "integralNumber": {
            "type": "boolean"
          },
          "floatingPointNumber": {
            "type": "boolean"
          },
          "array": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          },
          "null": {
            "type": "boolean"
          },
          "float": {
            "type": "boolean"
          },
          "short": {
            "type": "boolean"
          },
          "embeddedValue": {
            "type": "boolean"
          }
        }
      },
      "PushResult": {
        "type": "object",
        "properties": {
          "revision": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "StandardEdit": {
        "type": "object",
        "properties": {
          "operationId": {
            "type": "string",
            "format": "uuid"
          },
          "baseRevision": {
            "type": "integer",
            "format": "int64"
          },
          "reason": {
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/JsonNode"
          }
        }
      },
      "Change": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "baseRevision": {
            "type": "integer",
            "format": "int64"
          },
          "deleted": {
            "type": "boolean"
          },
          "payload": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/JsonNode"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "PushRequest": {
        "type": "object",
        "properties": {
          "operationId": {
            "type": "string",
            "format": "uuid"
          },
          "changes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Change"
            }
          },
          "catalogRevision": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
...
      "StandardArchive": {
        "type": "object",
        "properties": {
          "operationId": {
            "type": "string",
            "format": "uuid"
          },
          "baseRevision": {
            "type": "integer",
            "format": "int64"
          },
          "reason": {
            "type": "string"
          },
          "archived": {
            "type": "boolean"
          }
        }
      },
...
      "Record": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "revision": {
            "type": "integer",
            "format": "int64"
          },
          "deleted": {
            "type": "boolean"
          },
          "payload": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/JsonNode"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "Snapshot": {
        "type": "object",
        "properties": {
          "revision": {
            "type": "integer",
            "format": "int64"
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Record"
            }
          }
        }
      },
...
      "CatalogRecord": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "revision": {
            "type": "integer",
            "format": "int64"
          },
          "archived": {
            "type": "boolean"
          },
          "payload": {
            "$ref": "#/components/schemas/JsonNode"
          }
        }
      },
...
      "CatalogPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CatalogRecord"
            }
          },
          "hasMore": {
            "type": "boolean"
          },
          "offset": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}